+++ src/Bridge/Meilisearch/StoreFactory.php
@@ -1,0 +1,46 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\AI\Store\Bridge\Meilisearch;
+
+use Symfony\AI\Store\ManagedStoreInterface;
+use Symfony\AI\Store\StoreInterface;
+use Symfony\Component\HttpClient\HttpClient;
+use Symfony\Component\HttpClient\ScopingHttpClient;
+use Symfony\Contracts\HttpClient\HttpClientInterface;
+
+/**
+ * @author Guillaume Loulier <personal@guillaumeloulier.fr>
+ */
+final class StoreFactory
+{
+    public static function create(
+        string $indexName,
+        ?string $endpoint = null,
+        #[\SensitiveParameter] ?string $apiKey = null,
+        ?HttpClientInterface $httpClient = null,
+        string $embedder = 'default',
+        string $vectorFieldName = '_vectors',
+        int $embeddingsDimension = 1536,
+        float $semanticRatio = 1.0,
+    ): StoreInterface&ManagedStoreInterface {
+        if (null !== $endpoint) {
+            $defaultOptions = [];
+            if (null !== $apiKey) {
+                $defaultOptions['auth_bearer'] = $apiKey;
+            }
+
+            $httpClient = ScopingHttpClient::forBaseUri($httpClient ?? HttpClient::create(), $endpoint, $defaultOptions);
+        }
+
+        return new Store($httpClient, $indexName, $embedder, $vectorFieldName, $embeddingsDimension, $semanticRatio);
+    }
+}