+++ src/Bridge/Meilisearch/CHANGELOG.md
@@ -1,6 +1,14 @@
 CHANGELOG
 =========
 
+0.9
+---
+
+ * Introduce a `StoreFactory`
+ * [BC BREAK] Add support for `ScopingHttpClient` in `Store`
+ * [BC BREAK] The `endpointUrl` parameter for `Store` has been removed
+ * [BC BREAK] The `apiKey` parameter for `Store` has been removed
+
 0.1
 ---